+++ src/pages/task/taskTabs/buildBaronAndAnnotations/AnnotationTicketsTable/index.ts
@@ -1,0 +1,4 @@
+import AnnotationTicketsTable from "./AnnotationTicketsTable";
+import CustomCreatedTicketsTable from "./CustomCreatedTicketsTable";
+
+export { AnnotationTicketsTable, CustomCreatedTicketsTable };
